Revert "Revert "db연동 수정""
